+++ gframe/duelclient.h
@@ -29,7 +29,7 @@ private:
 	static char duel_client_read[0x2000];
 	static char duel_client_write[0x2000];
 	static bool is_closing;
-	static int select_hint;
+	static u64 select_hint;
 	static wchar_t event_string[256];
 	static mtrandom rnd;
 public:
